src/loader/json/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-auth
gunicorn
werkzeug
